Add files via upload
Assignment 4: Write a program to check if a given integer number is Positive, Negative or Zero. 	
Assignment 5: Write a program to check if a given integer number is odd or even.
Assignment 6: Write a program to check if the program has received command line arguments or not.
Assignment 7: Initialize two character variables in a program and display the characters in alphabetical order.
Assignment 8: Initialize a character variable in a program and print 'Alphabet' if the initialized value is an alphabet, print 'Digit' if the initialized value is a number, and print 'Special Character', if the initialized value is anything else.
Assignment 9: Write a program to accept gender ("Male" or "Female") and age from command line arguments and print the percentage of interest based on the given conditions.	
Assignment 10: Initialize a character variable with an alphabet in a program. If the character value is in lowercase, the output should be displayed in uppercase in the following format.
Assignment 11: Write a program to receive a color code from the user (an Alphabet). The program should then print the color name, based on the color code given. The following are the color codes and their corresponding color names.
R->Red, B->Blue, G->Green, O->Orange, Y->Yellow, W->White. If the color code provided by the user is not valid then print "Invalid Code". 
Assignment 12: Write a program to receive a number and print the corresponding month name.
Assignment 13: Write a program to print numbers from 1 to 10 in a single row with one tab space.
Assignment 14: Write a program to print even numbers between 23 and 57. Each number should be printed in a separate row.
Assignment 15: Write a program to check if a given number is prime or not.
Assignment 16: Write a program to print prime numbers between 10 and 99.
Assignment 17: Write a Java program to find if the given number is prime or not.
Assignment 18: Write a program to print the sum of all the digits of a given number.
Assignment 19: Write a program to print * in Floyd's format (using for and while loop).
Assignment 20: Write a program to reverse a given number and print.
Assignment 21: Write a Java program to find if the given number is palindrome or not.
Assignment 22: Write a program to print first 5 values which are divisible by 2, 3, and 5.